backend/foodgram/requirements.txt
django==2.2.19
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
djangorestframework-simplejwt==4.8.0
django-filter==21.1
python-dotenv~=0.20.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
pillow==9.2.0
reportlab==3.6.11
importlib-metadata==1.7.0
djoser==2.1.0